packages/ywpi/pyproject.toml
[build-system]
requires = ["setuptools"]
build-backend = "setuptools.build_meta"

[project]
name = "ywpi"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "aiochannel>=1.3.0",
    "grpcio-tools==1.64.0",
    "pydantic>=2.11.7",
    "watchfiles>=1.1.0",
]

[project.scripts]
ywpi = "ywpi.cli:main"
